move leaf optimization code into separate file + remove arrayencode::sum
